requirements.txt
flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
pyjwt>=2.9.0
flask-cors>=5.0.0